Minor fix reset script

+++ version3/scripts/reset.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
-killall python
-rm -rf .pidfile
-echo "Reset"
+pid=$(pid of python)
+if [ -z "$pid" ]; then
+    else "Reset not possible"
+    exit 1
+else
+    killall $pid
+    rm -rf .pidfile
+    echo "Reset"
+fi
